requirements.txt
pyside2
pyserial-asyncio